% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saraminR_Addin.R
\name{runsaramin}
\alias{runsaramin}
\title{Launch an RStudio addin which allows to schedule an Rscript interactively.}
\usage{
runsaramin()
}
\value{
the return of \code{\link[shiny]{runGadget}}
}
\description{
Launch an RStudio addin which allows to schedule an Rscript interactively.
}
\examples{
\dontrun{
cron_rstudioaddin()
}



}
